subscribe and unsubscribe to topics : properly working but can be improved in front (button "s'abonner" on topics page)

+++ back/src/main/java/com/openclassrooms/mddapi/service/SubscriptionService.java
@@ -1,6 +1,7 @@
 package com.openclassrooms.mddapi.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,6 +10,9 @@ import org.springframework.stereotype.Service;
 import com.openclassrooms.mddapi.dto.SubscriptionDto;
 import com.openclassrooms.mddapi.dto.TopicDto;
 import com.openclassrooms.mddapi.dto.UserDto;
+import com.openclassrooms.mddapi.exception.PostNotFoundException;
+import com.openclassrooms.mddapi.exception.SubscriptionAlreadyExistsException;
+import com.openclassrooms.mddapi.exception.SubscriptionNotFoundException;
 import com.openclassrooms.mddapi.exception.UserNotFoundException;
 import com.openclassrooms.mddapi.mapper.SubscriptionMapper;
 import com.openclassrooms.mddapi.mapper.UserMapper;
@@ -49,6 +53,7 @@ public class SubscriptionService implements ISubscriptionService {
      * @param SubscriptionDto the subscription that must be saved
      * @return SubscriptionDto the subscription saved
      * @throws UserNotFoundException 
+     * @throws SubscriptionNotFoundException 
      */
     // @Override
     // public SubscriptionDto create(SubscriptionDto subscriptionDto) {
@@ -65,12 +70,21 @@ public class SubscriptionService implements ISubscriptionService {
     // }
 
     @Override
-    public SubscriptionDto create(Long topicId, String email) throws UserNotFoundException {
+    public SubscriptionDto create(Long topicId, String email) throws UserNotFoundException, SubscriptionAlreadyExistsException {
 
         log.debug("Trying to save the subscription by user with email {} to topic with id {}", email, topicId);
 
         // Chercher le user 
         UserDto userDto = userService.findByEmail(email);
+
+        /// Chercher si la souscription existe déjà 
+
+        Optional<Subscription> optionalSubscription = subscriptionRepository.findByUserIdAndTopicId(userDto.getId(), topicId);
+            
+        if(optionalSubscription.isPresent()) {
+            log.error("Subscription by user with email {} to topic with id {} already exists", email, topicId);
+            throw new SubscriptionAlreadyExistsException("Subscription by user with email " +  email + " to topic with id " + topicId + " already exists");
+        }
 
         // Chercher le topic 
         TopicDto topicDto = topicService.getTopicById(topicId);
@@ -90,16 +104,22 @@ public class SubscriptionService implements ISubscriptionService {
     }
 
 
-
     @Override
-    public void delete(Long topicId, String email) throws UserNotFoundException {
+    public void delete(Long topicId, String email) throws UserNotFoundException, SubscriptionNotFoundException {
 
         log.debug("Trying to delete the subscription by user with email {} to topic with id {}", email, topicId);
 
         // Chercher le user Id
         Long userId = userService.findByEmail(email).getId();
 
-        Subscription subscriptionToDelete = subscriptionRepository.findByUserIdAndTopicId(userId, topicId);
+        Optional<Subscription> optionalSubscription = subscriptionRepository.findByUserIdAndTopicId(userId, topicId);
+
+        if (!optionalSubscription.isPresent()) {
+            log.error("Subscription by user with email {} to topic with id {} doesn't exist", email, topicId);
+            throw new SubscriptionNotFoundException("Subscription by user with email " + email + " to topic with id " + topicId + " doesn't exist");
+        }
+
+        Subscription subscriptionToDelete = optionalSubscription.get();
 
         subscriptionRepository.delete(subscriptionToDelete);
